+++ NEWS.md
@@ -1,5 +1,7 @@
 # pw (development version)
 
+* `pw_show_report()` can be called to open the test report
+
 * `pw_codegen()` can be called to launch the code generator
 
 * `pw_test()` stops if the playwright skeleton is missing